Now connections keep state of what they haven't considered yet.
By associating each detour entry with a timestamp, each connection tracks what
the most recent entry they looked at was, and the subflow worker only considers
entries newer than that. This way, we can safely and repeatedly call
create_subflow_worker. Additionally, on detour_add we iterate over every
connection in the namespace and wake it (for NAT only if it is relevant). With
this we can be certain to dynamically add detours, but also avoid priv_list and
the troubles we had with trying to cleanup path manager data.